+++ source/KDTree.cpp
@@ -22,7 +22,7 @@ void KDTree::build(const std::vector<std::vector<double>>& points, const std::ve
     if (points.empty()) return;
     dimension = points[0].size();
     std::vector<std::pair<std::vector<double>, int>> point_labels(points.size());
-    for (size_t i = 0; i < points.size(); ++i)
+    for (unsigned int i = 0; i < points.size(); ++i)
     {
         point_labels[i] = std::make_pair(points[i], labels[i]);
     }
@@ -45,7 +45,7 @@ KDNode* KDTree::buildTree(std::vector<std::pair<std::vector<double>, int> >& poi
     if (points.empty()) return nullptr;
 
     int axis = depth % dimension;
-    size_t median = points.size() / 2;
+    unsigned int median = points.size() / 2;
 
     // Use the functor as the comparison
     std::nth_element(points.begin(), points.begin() + median, points.end(), ComparePairs(axis));
@@ -100,7 +100,7 @@ void KDTree::nearest(KDNode* root, const std::vector<double>& target, KDNode*& b
 double KDTree::calculateDistance(const std::vector<double>& a, const std::vector<double>& b) const
 {
     double sum = 0.0;
-    for (size_t i = 0; i < a.size(); ++i)
+    for (unsigned int i = 0; i < a.size(); ++i)
     {
         double diff = a[i] - b[i];
         sum += diff * diff;
