+++ source/KNNClassifier.cpp
@@ -59,14 +59,14 @@ void KNNClassifier::fit(std::string abs_filepath_to_features, std::string abs_fi
     }
     pVec->fit(abs_filepath_to_features, abs_filepath_to_labels);
 
-    size_t num_features = pVec->word_array.size();
+    unsigned int num_features = pVec->word_array.size();
     std::vector<std::tr1::shared_ptr<Sentence> > sentences = pVec->sentences;
     training_features.clear();
     training_labels.clear();
 
     std::ifstream label_file(abs_filepath_to_labels.c_str());
     int label;
-    for (size_t i = 0; i < sentences.size(); ++i)
+    for (unsigned int i = 0; i < sentences.size(); ++i)
     {
         std::vector<double> features;
 		const std::tr1::unordered_map<int, double>& sentence_map = sentences[i]->sentence_map;
@@ -104,7 +104,7 @@ void KNNClassifier::predict(std::string abs_filepath_to_features, std::string ab
     #ifdef BENCHMARK
     double sumduration = 0.0;
     double sumstrlen = 0.0;
-    size_t num_rows = 0;
+    unsigned int num_rows = 0;
     #endif
 
     while (getline(in, feature_input))
@@ -149,7 +149,7 @@ Prediction KNNClassifier::predict(std::string sentence, bool preprocess)
     // Calculate probability
     double total_distance = 0.0;
     std::vector<double> closest_distances = kd_tree.getClosestDistances(feature_vector, k);
-    for (size_t i = 0; i < closest_distances.size(); ++i)
+    for (unsigned int i = 0; i < closest_distances.size(); ++i)
     {
         total_distance += closest_distances[i];
     }
@@ -167,7 +167,7 @@ int KNNClassifier::getLabel(const std::vector<double>& features) const
 double KNNClassifier::calculateDistance(const std::vector<double>& a, const std::vector<double>& b) const
 {
     double sum = 0.0;
-    for (size_t i = 0; i < a.size(); ++i)
+    for (unsigned int i = 0; i < a.size(); ++i)
     {
         double diff = a[i] - b[i];
         sum += diff * diff;
@@ -186,16 +186,16 @@ void KNNClassifier::save(const std::string& filename) const
 
     pVec->save(outFile);
 
-    size_t training_features_size = training_features.size();
+    unsigned int training_features_size = training_features.size();
     outFile.write(reinterpret_cast<const char*>(&training_features_size), sizeof(training_features_size));
-    for (size_t i = 0; i < training_features_size; ++i)
+    for (unsigned int i = 0; i < training_features_size; ++i)
     {
-        size_t features_size = training_features[i].size();
+        unsigned int features_size = training_features[i].size();
         outFile.write(reinterpret_cast<const char*>(&features_size), sizeof(features_size));
         outFile.write(reinterpret_cast<const char*>(&training_features[i][0]), features_size * sizeof(int));
     }
 
-    size_t training_labels_size = training_labels.size();
+    unsigned int training_labels_size = training_labels.size();
     outFile.write(reinterpret_cast<const char*>(&training_labels_size), sizeof(training_labels_size));
     outFile.write(reinterpret_cast<const char*>(&training_labels[0]), training_labels_size * sizeof(int));
 
@@ -213,18 +213,18 @@ void KNNClassifier::load(const std::string& filename)
 
     pVec->load(inFile);
 
-    size_t training_features_size;
+    unsigned int training_features_size;
     inFile.read(reinterpret_cast<char*>(&training_features_size), sizeof(training_features_size));
     training_features.resize(training_features_size);
-    for (size_t i = 0; i < training_features_size; ++i)
+    for (unsigned int i = 0; i < training_features_size; ++i)
     {
-        size_t features_size;
+        unsigned int features_size;
         inFile.read(reinterpret_cast<char*>(&features_size), sizeof(features_size));
         training_features[i].resize(features_size);
         inFile.read(reinterpret_cast<char*>(&training_features[i][0]), features_size * sizeof(int));
     }
 
-    size_t training_labels_size;
+    unsigned int training_labels_size;
     inFile.read(reinterpret_cast<char*>(&training_labels_size), sizeof(training_labels_size));
     training_labels.resize(training_labels_size);
     inFile.read(reinterpret_cast<char*>(&training_labels[0]), training_labels_size * sizeof(int));
